rename migration 0037 to what it acutally does
