+++ database/entity/0037-drop_user_setting_table/User.ts
